Use fmap images in dwidenoise and unring
By concatenating the images within the fmap/ directory with the input DWI, MP-PCA denoising and Gibbs ringing removal can additionally be applied to these images. They are later separated from the input DWI volumes before running dwipreproc, so that those volumes in fmap are not included in the output of dwipreproc.