Command-line options: Split container and standslone usage
Rather than providing both single-dash and double-dash usage strings for each command-line option (for both compatibility with OpenNeuro and consistency with MRtrix3), these changes completely split the construction of command-line options depending on whether the script is running within a container or as a stand-alone script (against the MRtrix3 Python libraries). In the case of the former, the standard options provided by the MRtrix3 libraries are erased, and replaced with only relevant options, all with double-dash usage. For the latter, the standard MRtrix3 interface is retained: single-dash option strings and standard options that are common across all scripts.
First discussed in #12.