Construct symbolic links for FreeSurfer
It appears as though recon-all searches for the subjects 'fsaverage', 'lh.EC_average' and 'rh.EC_average' within the same directory as specified with the -sd command-line option, rather than the SUBJECTS_DIR environment variable.
This change constructs symbolic links within the overridden subject directory so that recon-all finds the files it searches for.