New approach to T1 brain extraction
If available, use a combination of N4 and ROBEX to perform bias field correction and brain extraction. If these two commands are not available, revert to using fsl_anat to perform these two tasks.
Also, regardless of the commands used here, set the strides of the output images to be identical to the input T1 image.