Support group-level analysis with --output_verbosity 1
Because group-level analysis relies on an FA image for registration and a mean b=0 image for obtaining a median white matter reference value, such processing wouuld fail if participant-level analysis was run with --output_verbosity 1, since the pre-processed DWI would not be written to the output directory in this instance.
This change ensures that even for --output_verbosity 1, a mean b=0 image and an FA image are generated during participant-level processing, to then be read during group-level processing.
In addition, the lookup table file for the parcellation is written to the output target directory (not the subject-specific directory), as this provides convenience for navigating the command output.