No longer concatenate dwi/ and fmap/ data
Previously, the script would concatenate all data in the dwi/ and fmap/ directories together prior to running dwidenoise / mrdegibbs. This placed limitations on such images being defined on a common grid, and it may not be ideal from the perspective of denoising to be concatenating data with different phase encoding directions beforehand. With this change, DWI series are denoised independently, and all DWI and fmap/ series have Gibbs ringing removal applied independently. Finally, if it is not possible to directly concatenate all fmap/ data together prior to dwipreproc, all fmap/ images will be resampled onto the voxel grid of the DWIs in preparation for dwipreproc.