Fix complex DWI import PE
On import of complex DWI data, the import of JSON data needs to occur in the same mrconvert call as where the NIfTI image is loaded. If an mrcalc call is first performed to combine magnitude and phase components, and then mrconvert is used to bring in the JSON information, then the phase encoding direction will not be appropriately rotated when MRtrix3 re-arranges the spatial axes to RAS.
Also sort input DWIs and fmap/ images, so that if volumes have been extracted from the input DWIs and fed to fmap/, then the first volume input to topup is more likely to correspond to the first input volume to eddy.
Also retain dwifslpreproc scratch directory if MRtrix3_connectome is being run with -nocleanup.