participant2: Check output paths at commencement
Major structural change to how output files from participant2-level analysis are managed. In particular initialising this information at the start of the function allows for checking for pre-existing files. This also has the side-effect of making the structure of the outputs of this function a little easier to follow / understand.