Fix and generalise eddy outlier crash detection
It appears as though the error message generated by eddy in the case where it can't find a good reference direction with no outliers for a particular b-value shell differs between different versions of the FSL software. This change will hopefully catch this issue across multiple versions.
It also fixes a syntax error in the catching of the exception from the unsuccessful eddy execution.