participant: Change ANTs template registration
Due to issues observed with the outputs of the WarpImageMultiTransform command, where large areas of the brain would be cropped out, this change completely replaces the ANTs command calls with examples from a different manuscript from the ANTs developers. This includes a change from WarpImageMultiTransform command, which is apparently deprecated, to the antsApplyTransforms command.