Containers: Remove core dependency version specificity
While installation of specific versions of core system dependencies were added for reproducibility of container recipes, this quickly led to specific dependencies being missing from the repositories. Moreover, the principal driver for this specificity was for the dependencies for EddyQC, but this is now included in the FSL installation and does not need to be installed from the git repository. Therefore these core dependencies are being reverted to install the latest version of each.