Temporary fix for FreeSurfer
Runs FreeSurfer on a brain-extracted T1-weighted image if provided with one, or on the original raw T1-weighted image if not. However this will fail to address the case where there is a pre-processed T1-weighted image that is skull-stripped but the original unprocessed T1-weighted image could nevertheless have been recovered.