dwifslpreproc: Use --data_is_shelled on unknown failure
An instance was discovered where FSL eddy fails due to judging the DWI data to not be a shelled acquisition, but the contents that it writes to stdout is some form of binary data rather than the expected error meswsage.
With this change, if eddy fails due to an unknwon reason, but it has not yet been attempted with this option included, that second attempt will be made, even though the corresponding error message has not yet been encountered.