Change MRtrix3 version for dwicat update
In d4b25b38095c3177e8e65f865721a056c03bd6a5, the version of MRtrix3 was updated to introduce a feature addition to dwicat. However, this involved changing that dependency to the development branch of MRtrix3, which contains changes over and above just the enhancement of dwicat.
With this commit, the version of MRtrix3 installed is genuinely 3.0.4, with some additional unrelated hotfixes since that tag, plus the cherry-picked enhancement to dwicat, such that no other major changes to MRtrix3 are introduced.