Fix use of FSL for derivation of brain mask only
Where the user provides as input a derivatives directory that only provides a pre-processed T1-weighted image and not a corresponding mask, FSL's fsl_anat command may produce images for which the field of view in specifically the inferior portion of the image has been reduced. This would otherwise lead to downstream problems where the mask extracted from this directory and the input pre-processed T1-weighted image are expected to reside on the same voxel grid, initially by the mrhistmatch command. This fix resamples the FSL-derived brain mask back onto the full voxel grid of the input pre-processed T1-weighted image.